Add files via upload
As a graduated master student in Sweden, I've chosen to analyze data from IKEA, one of the world's largest Swedish furniture companies. IKEA is renowned not only for its affordable and stylish furniture but also for its commitment to innovation and the application of AI/ML in various aspects of its operations. This project aims to explore IKEA's product data, focusing on pricing strategies, product categories, and potential discounts.